feat(Gameboard): initialize deck and display its cards
A deck containing cards data has been initialized in the GameBoard
component. It is temporarily defined as just a local variable
initialized with some sample data (a placeholder image is used as card
face), to test the rendering of the component.

Later on, the deck will be handled differently and the card faces will
be grabbed through API calls.

Then, the cards of the table (tableCards) have been initialized with
such full deck. Later on, only a part of the deck will be shown in the
table.