feat: add helper functions to fetch the data for each card
The initCardData() helper function initialize the data (id, url) for a
new card, fetching an image using the provided api, converting it to a
blob object and getting its url; a unique random id is also generated.
You can pass a signal to abort the fetching when needed.

The fetchCardsData() helper function fetches the data for a given number
of cards and uses the initCardData() function. Again, you can pass a
signal to abort the fetching when needed.