feat(useDeck): add useDeck(desiredDeckSize) custom hook to handle the deck
The useDeck(desiredDeckSize) custom hook has been defined to handle the
deck array containing the info of the possible cards to use, and making
it have desiredDeckSize size.
The deck is returned by the hook, however, is does not necessarily have
desiredDeckSize size initially, because such deck is filled
asyncronously to deckSize by useDeck hook.
Also, the returned deck should be considered immutabile (it is a state
inside the hook).
Passing a larger deckSize updates the deck with new cards. A render is
always triggered when the deck is updated.

The useDeck has been used within the GameBoard component to handle the
deck in thre.
A deckSize state initialized to 4 has been defined in there and used as
desiredDeckSize input for the useDeck hook.
A click callback on the component which increases deckSize by 1 has been
temporarily defined to test the hook, too.

The hook has been tested against some possible race conditions occurring
in development (with StrictMode enabled) or when multiple size updates
are issued quickly, also enabling low-end mobile device on DevTools, to
simulate slow fetching.