feat(GameBoard): update click callback to select cards on click
The useSelectedCards() custom hook has been used to set a card as
selected and to check that condition.

Then, the click callback on the .gameboard element has been modified to
select cards on click.

A single event listener is used exploiting event delegation. When a
click is captured, first is is checked whether there is a card in the
clicked point.

In that case, if the associated card id has already been selected, an
alert is shown. Otherwise, the card is set as selected and the game
temporarily proceeds as before (incrementing the deck size).

For development only, a style has been applied to the selected cards, to
highlight them.