feat(GameView): add score state to GameView component
The score is defined as a state variable within the GameView component.
Although it is not rendered in the page yet, it will be displayed in a
header element outside the Gameboard view, but within the GameView
element.

The score is initialized to 0 and incremented by the Gameboard component
whenever the player hits a non-selected card.