feat(App): add bestScore state to App component
The bestScore is defined as a state variable within the App component.
Although the restart game functionality is not implemented yet, this
best score is independent from the instance of the game played (and
should be retained between games), and so it is instantiated outside the
GameView component.

In the future, such value will also be saved in LocalStorage.

bestScore and SetBestScore are passed to the GameviewComponent, where
the incrementScore(delta) function is defined which increments
the current score by delta, and also checks whether the bestScore is
improved or not.

Function incrementScore is now passed to the GameBoard component instead
of setState.

Both score and bestScore are still only printed to the console, and not
displayed on the page yet.

bestScore is initialized to 0. However, to test whether the score /
bestScore update mechanism works, it can be temporarily set to some
non-zero value.