feat(GameView): add isNewBestScore ref to GameView component
isNewBestScore is a ref which keeps track of whether the current game
has improved the app bestScore.

Is is set to true within incrementScore(delta) function, just before the
setBestScore is called (which triggers a rerender).

Score, bestScore and isNewBestScore are still only printed to the
console, and not displayed on the page, yet.