refactor(useDeck): make deck a Map object where each item is indexed by its indexed
Previously deck was an array of objects, each one representing a card.
Now, this is a Map object where each of such items are indexed by their
id.

This will be used to simplify the handling of the cards shown in the
table, which is just a part of the deck (see next commit).