feat(useTableCards): add useTableCards custom hook
This custom hook can be used to manage the cards shown in the table
during the gameplay.

You must specify the desired size of the table, the fraction of selected
cards to be included in the table (provided that there are enough...
otherwise, all the selected cards are shown in the table), the deck Map,
the isSelectedCard() function to check whether a given card has already
been selected, and the total number of selected cards.

Within this custom hook, the tableCards array has been defined as a
React state, which is initialized using the local getTableCards()
function, which computes the number of selected and unselected cards to
be included in the table, and randomly choose them form the deck; the
cards of the table are eventually shuffled again.

The custom hook returns two things:
- the tableCards state (which is to be considered immutable);
- the refreshTableCards() function which generates a new combination of
  cards in the table, again using the getTableCards() function.

This custom hook has been used in the GameBoard component. Note that the
returned JSX has been modified to the new stucture of tableCards (now it
is an array of cards'ids).