feat(): add GameOverView component
The component creates the view to be shown when the game ends.

It shows a "game over" message, along with the final score and possibly
a message telling whether a new score has been achieved.

The view has been instantiated and temporarily set as default view in
the App component, for testing.