feat(GameView): add setGameOverViewCallback input prop to GameView component
setGameOverViewCallback is a callback that can be called when the game
is over, to show the GameOverView component.

It has two inputs: the final score and the isNewBestscore flag, and is
used to define the gameOverCallback function to be passed to the
GameBoard component.

In the App component, the default view has been set back to "game", and
the setGameOverViewCallback prop passed to the GameView component has
been properly set to change the view to the GameOverView one.

Note that the final score and the isNewBestscore flag are passed to the
currentView object (data field) through a callback, and not shifting
those state up from the GameView component to the App component: this is
because those states are local and specific to the game instance, and
not to the whole app, and reset automatically when a new game is started
(the GameView component is mounted).