refactor(Card,GameCard)!: refactor code for Card components
Now the Game component is a generic container repesenting a generic
Card. The actual content is passed through the children props.

This allows to reuse the code for, eg, showing the game-over message as
a card, and  allow consistent and imporved styling for the cards.

Then, in the GameBoard component, the (new) GameCard is used, which
internally uses the Card component and specifies an image as a child.

Also, in the GaveOverComponent, the game over messages are wrapped in a
Card component, too, to make them resemble a card.

BREAKING CHANGE: the former Card component has been renamed to GameCard
component and uses the new Card component.