refactor(Card): refactor Card component to a different structure
The component now has a front face and a rear face div. The children
props is applied as children of the front face.

This will be used to handle the animations on the cards.

Note that in GameCard, the image has been given a different class name,
to avoid conflicting names.

The corresponding styles have been adapted, too.