feat(GameBoard): minor adjustments for the ready state effect handling in GameBoard component
In GameBoard component, some delay has been applied when setting the
ready state, based on whether the fetching is the first one or one
occurring during the gameplay.

A different ready class name has been applied for the first fetch
("ready-init"), too, in order to be able to apply styles specific to the
first render.