style(Card): add flip animations to Card components
The type of animation depends on the class applied to the Card
component, eg, it is different between the .game-card and the
.game-over-card.

The backface visibility of the card faces has been set to hidden, and
the .card-rear face of the .card has been rotaded along the y axis by
180 degrees. When the .card element is rotated in the 3d space, the
front face gets hidden and the rear face becomes visible.

A perspective is also applied to the parent of the .card element, to
give the effect of a card lifting up while flipping.

This is based on:  https://auroratide.com/posts/realistic-flip-animation/