fix(Gameboard): set clickCallback click callback Gameboard component to trigger when a .game-card ancestor to the e.target is detected
The e.target.closest(".game-card") method is used.

Before this change, the clickCallback was not ignored only when
.game-card-img was the e.target, ie, the top most element. If there are
other elements above it, the card is not detected correctly.

The data-id attribute of .game-card-img is now assigned to the
.card.game-card div. To do that, the customAttribute prop is defined for
the Card component, which you can use to pass an object of custom
attributes.