feat(GameBoard): add initialDeckSize, incrementDeckSize, tableSize, selectedCardsFractInTable props to GameBoard component
The initial value of deckSize and its increment are passed to the
component through the initialDeckSize and incrementDeckSize props.

The number of cards in the table and the fraction of selected cards in
there are are passed to through the tableSize and
selectedCardsFractInTable props.

All these props have been given default values, too.

This will allow to setup different levels for the game.