feat(HomeView): setup HomeView component
The HomeView view component will be used to select a game level to play.

Currently, only a button for each level is instantiated, with a simple
label; however, no click event listener is set, yet. The levels info are
passed to this component using props.

The HomeView view component is instantiated in the App component when
the currentView.name state is set to "Home". Levels info are retrieved
from  ./data/levelsSettings.json.