feat(HomeView): add setPlayGameCallback prop to HomeView component
The setPlayGameCallback(level) prop is a function that setup the
callback to be used when clicking on the button associated with a given
level.

In the App component, where the HomeView component is instantiated, the
setPlayGameCallback function is initialized (it simply updates the
currentView state, setting the view as "game" and passing the level
through the data property) and passed to it.

The gameSettings prop passed to the GameView component in the App
component is now currentView.data.level.