feat(Title): add onClickCallback on Title component
The optional onClick callback can be setup for the Title component to
setup an action.

In practice, we want to return to the home page when clicking on that
element. Then, in the App component, the setHomeViewCallback function is
defined which sets the current view to the HomeView one. The function is
then passed as prop to the GameOverView and GameView components where it
is forwarded to the Title component.

Note that the backBtnCallback passed to GameOverView component now is
set to show the HomeView component, too.