feat(GameBoard): add props to set the initial tableSize and its increment in GameBoard component + add Incremental level
This allows to have a new level (Incremental, as defined in
levelsSettings.json), where you start from three cards and add one card
at each turn.

Then, tableSize is now a state in GameBoard component initialized based
on the initialTableSize prop, and the value is properly incremented, if
needed, in the clickCallback function.

Note that the increments to be used to update deckSize and tableSize are
now Refs defined on component mount, clipped to avoid negative
increments, so such data are "validated" to avoid reducing those sizes.

Note that when incrementing the deck size, you have to consider also the
increment on the table, and also you have to initialize the deckSize
with the maximum between initialDeckSize and initialTableSize, to avoid
being out of unselected cards to be used in there.