fix(useTableCards): fix clip of numOfSelectedCardsInTable in useTableCards custom hook
Limit numOfSelectedCardsInTable to tableSize - 1, to leave at least one
not selected item and, provided that there is at least one selected
card, make numOfSelectedCardsInTable at least 1, to show a selected card
in the table.

This allows to allow both the player to continue playing and lose the
game.

Recall that numOfSelectedCardsInTable is also limited by the number of
selected cards.