feat(GameBoard): add possibility to have an increasing selectedCardsFractInTable
selectedCardsFractInTable (denoted y(k) below) roughly represent the
desired fraction of selected cards shown on the table, provided that
there are enough of them (if this is not the case, all the selected
cards are shown).

Now it evolves according to initialSelectedCardsFractInTable,
limitSelectedCardsFractInTable, and aSelectedCardsFractInTable, which
are provided through props.

In particular, selectedCardsFractInTable now starts from
initialSelectedCardsFractInTable and increases over time converging
to limitSelectedCardsFractInTable with a speed depending on
aSelectedCardsFractInTable, evolving according to:

   y(k) = y0 + (b - y0) * (1 - a^k),

where:
  y(0) = y0
  y(k) -> b for k -> +inf

  a: aSelectedCardsFractInTable, 0<a<=1
  b: limitSelectedCardsFractInTable
 y0: initialSelectedCardsFractInTable

that is:

   y(k+1) = (1-a)*b + a * y(k)

Constraints:
  0 <= initialSelectedCardsFractInTable <= limitSelectedCardsFractInTable <= 1
  0 < aSelectedCardsFractInTable < 1

If you want to keep the value constant, just use
limitSelectedCardsFractInTable = initialSelectedCardsFractInTable and
aSelectedCardsFractInTable = 1, which are the default values for such
props, ie, just provide initialSelectedCardsFractInTable as prop.

The "incremental" difficulty game level has been upgraded to have a
variable selectedCardsFractInTable (for the other levels this is still
constant). In particular, it starts from 0.3 and slowly approaches 1 as
the game goes on, to make the difficuly increase.