feat(useCardsPerRowAndCol): add useCardsPerRowAndCol custom hook
The useCardsPerRowAndCol custom hook is used to set two CSS variables
(defined by string parameters cardsPerRowCssVar and cardsPerColCssVar),
which define the number of rows and columns in a grid to accommodate
numOfCards squared cards, ensuring they fit within a given HTML element.
The custom hook returns a ref which is to be used as ref in the element
you want to fit the cards in.

This custom hook has been used in the HomeView and GameBoard components,
to set the --home-level-cards-per-row, --home-level-cards-per-col,
--gameboard-cards-per-row, --gameboard-cards-per-col CSS variables,
based on the number of levels (gameLevels.length) and tableSize,
respectively.

Then, these CSS variables are used to properly set the --cards-per-row
and --cards-per-col CSS variables, depending on the
.gameboard.cards-container or .home-view .cards-container selector,
respectively.

Using the ".cards-container > .card"  selector, the --card-size CSS
variable has been re-defined, based on --cards-per-row and
--cards-per-col, so that the cards fit in there.

Note that .cards-container remains a flex container with wrapping, so
that the cards per row are not necessarily equal to --cards-per-row.
Consider transforming it to a grid container.