fix(GameBoard): update tableSize and selectedCardsFractInTable after the fetching is done
Before this fix, the tableSize and selectedCardsFractInTable were
updated just after a card was selected. The effect of this was shown
after the fetching of new deck cards was completed, after
delayFetchingToReadyInMs ms, when the deck was updated.

With the latest introduction of useCardsPerRowAndCol, this possibly
caused the card grid distribution to change before the deck was updated
(eg, in the Incremental difficulty level).

To avoid this, the updating of tableSize and selectedCardsFractInTable
has been moved after the deck has been updated. To trigger the refresh
of the table cards, a temprary game state "update-table-settings" has
been inserted before the setting the "ready" state.

Recall that this is necessary because refreshTableCards does not trigger
at each render, but on demand.